+++ lib/cubit/broadcast.freezed.dart
@@ -1,0 +1,379 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
+
+part of 'broadcast.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
+/// @nodoc
+mixin _$BroadcastState {
+  bool get broadcasting => throw _privateConstructorUsedError;
+  String get errBroadcasting => throw _privateConstructorUsedError;
+  String get psbt => throw _privateConstructorUsedError;
+  String get hex => throw _privateConstructorUsedError;
+  String get txId => throw _privateConstructorUsedError;
+  String get errFileImport => throw _privateConstructorUsedError;
+  bool get clearPsbt => throw _privateConstructorUsedError;
+  String? get importedPsbtPath => throw _privateConstructorUsedError;
+  String? get importedPsbtfileName => throw _privateConstructorUsedError;
+  String? get importedHexPath => throw _privateConstructorUsedError;
+  String? get importedHexfileName => throw _privateConstructorUsedError;
+
+  @JsonKey(ignore: true)
+  $BroadcastStateCopyWith<BroadcastState> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $BroadcastStateCopyWith<$Res> {
+  factory $BroadcastStateCopyWith(
+          BroadcastState value, $Res Function(BroadcastState) then) =
+      _$BroadcastStateCopyWithImpl<$Res>;
+  $Res call(
+      {bool broadcasting,
+      String errBroadcasting,
+      String psbt,
+      String hex,
+      String txId,
+      String errFileImport,
+      bool clearPsbt,
+      String? importedPsbtPath,
+      String? importedPsbtfileName,
+      String? importedHexPath,
+      String? importedHexfileName});
+}
+
+/// @nodoc
+class _$BroadcastStateCopyWithImpl<$Res>
+    implements $BroadcastStateCopyWith<$Res> {
+  _$BroadcastStateCopyWithImpl(this._value, this._then);
+
+  final BroadcastState _value;
+  // ignore: unused_field
+  final $Res Function(BroadcastState) _then;
+
+  @override
+  $Res call({
+    Object? broadcasting = freezed,
+    Object? errBroadcasting = freezed,
+    Object? psbt = freezed,
+    Object? hex = freezed,
+    Object? txId = freezed,
+    Object? errFileImport = freezed,
+    Object? clearPsbt = freezed,
+    Object? importedPsbtPath = freezed,
+    Object? importedPsbtfileName = freezed,
+    Object? importedHexPath = freezed,
+    Object? importedHexfileName = freezed,
+  }) {
+    return _then(_value.copyWith(
+      broadcasting: broadcasting == freezed
+          ? _value.broadcasting
+          : broadcasting // ignore: cast_nullable_to_non_nullable
+              as bool,
+      errBroadcasting: errBroadcasting == freezed
+          ? _value.errBroadcasting
+          : errBroadcasting // ignore: cast_nullable_to_non_nullable
+              as String,
+      psbt: psbt == freezed
+          ? _value.psbt
+          : psbt // ignore: cast_nullable_to_non_nullable
+              as String,
+      hex: hex == freezed
+          ? _value.hex
+          : hex // ignore: cast_nullable_to_non_nullable
+              as String,
+      txId: txId == freezed
+          ? _value.txId
+          : txId // ignore: cast_nullable_to_non_nullable
+              as String,
+      errFileImport: errFileImport == freezed
+          ? _value.errFileImport
+          : errFileImport // ignore: cast_nullable_to_non_nullable
+              as String,
+      clearPsbt: clearPsbt == freezed
+          ? _value.clearPsbt
+          : clearPsbt // ignore: cast_nullable_to_non_nullable
+              as bool,
+      importedPsbtPath: importedPsbtPath == freezed
+          ? _value.importedPsbtPath
+          : importedPsbtPath // ignore: cast_nullable_to_non_nullable
+              as String?,
+      importedPsbtfileName: importedPsbtfileName == freezed
+          ? _value.importedPsbtfileName
+          : importedPsbtfileName // ignore: cast_nullable_to_non_nullable
+              as String?,
+      importedHexPath: importedHexPath == freezed
+          ? _value.importedHexPath
+          : importedHexPath // ignore: cast_nullable_to_non_nullable
+              as String?,
+      importedHexfileName: importedHexfileName == freezed
+          ? _value.importedHexfileName
+          : importedHexfileName // ignore: cast_nullable_to_non_nullable
+              as String?,
+    ));
+  }
+}
+
+/// @nodoc
+abstract class _$$_BroadcastStateCopyWith<$Res>
+    implements $BroadcastStateCopyWith<$Res> {
+  factory _$$_BroadcastStateCopyWith(
+          _$_BroadcastState value, $Res Function(_$_BroadcastState) then) =
+      __$$_BroadcastStateCopyWithImpl<$Res>;
+  @override
+  $Res call(
+      {bool broadcasting,
+      String errBroadcasting,
+      String psbt,
+      String hex,
+      String txId,
+      String errFileImport,
+      bool clearPsbt,
+      String? importedPsbtPath,
+      String? importedPsbtfileName,
+      String? importedHexPath,
+      String? importedHexfileName});
+}
+
+/// @nodoc
+class __$$_BroadcastStateCopyWithImpl<$Res>
+    extends _$BroadcastStateCopyWithImpl<$Res>
+    implements _$$_BroadcastStateCopyWith<$Res> {
+  __$$_BroadcastStateCopyWithImpl(
+      _$_BroadcastState _value, $Res Function(_$_BroadcastState) _then)
+      : super(_value, (v) => _then(v as _$_BroadcastState));
+
+  @override
+  _$_BroadcastState get _value => super._value as _$_BroadcastState;
+
+  @override
+  $Res call({
+    Object? broadcasting = freezed,
+    Object? errBroadcasting = freezed,
+    Object? psbt = freezed,
+    Object? hex = freezed,
+    Object? txId = freezed,
+    Object? errFileImport = freezed,
+    Object? clearPsbt = freezed,
+    Object? importedPsbtPath = freezed,
+    Object? importedPsbtfileName = freezed,
+    Object? importedHexPath = freezed,
+    Object? importedHexfileName = freezed,
+  }) {
+    return _then(_$_BroadcastState(
+      broadcasting: broadcasting == freezed
+          ? _value.broadcasting
+          : broadcasting // ignore: cast_nullable_to_non_nullable
+              as bool,
+      errBroadcasting: errBroadcasting == freezed
+          ? _value.errBroadcasting
+          : errBroadcasting // ignore: cast_nullable_to_non_nullable
+              as String,
+      psbt: psbt == freezed
+          ? _value.psbt
+          : psbt // ignore: cast_nullable_to_non_nullable
+              as String,
+      hex: hex == freezed
+          ? _value.hex
+          : hex // ignore: cast_nullable_to_non_nullable
+              as String,
+      txId: txId == freezed
+          ? _value.txId
+          : txId // ignore: cast_nullable_to_non_nullable
+              as String,
+      errFileImport: errFileImport == freezed
+          ? _value.errFileImport
+          : errFileImport // ignore: cast_nullable_to_non_nullable
+              as String,
+      clearPsbt: clearPsbt == freezed
+          ? _value.clearPsbt
+          : clearPsbt // ignore: cast_nullable_to_non_nullable
+              as bool,
+      importedPsbtPath: importedPsbtPath == freezed
+          ? _value.importedPsbtPath
+          : importedPsbtPath // ignore: cast_nullable_to_non_nullable
+              as String?,
+      importedPsbtfileName: importedPsbtfileName == freezed
+          ? _value.importedPsbtfileName
+          : importedPsbtfileName // ignore: cast_nullable_to_non_nullable
+              as String?,
+      importedHexPath: importedHexPath == freezed
+          ? _value.importedHexPath
+          : importedHexPath // ignore: cast_nullable_to_non_nullable
+              as String?,
+      importedHexfileName: importedHexfileName == freezed
+          ? _value.importedHexfileName
+          : importedHexfileName // ignore: cast_nullable_to_non_nullable
+              as String?,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$_BroadcastState extends _BroadcastState with DiagnosticableTreeMixin {
+  const _$_BroadcastState(
+      {this.broadcasting = false,
+      this.errBroadcasting = '',
+      this.psbt = '',
+      this.hex = '',
+      this.txId = '',
+      this.errFileImport = '',
+      this.clearPsbt = false,
+      this.importedPsbtPath,
+      this.importedPsbtfileName,
+      this.importedHexPath,
+      this.importedHexfileName})
+      : super._();
+
+  @override
+  @JsonKey()
+  final bool broadcasting;
+  @override
+  @JsonKey()
+  final String errBroadcasting;
+  @override
+  @JsonKey()
+  final String psbt;
+  @override
+  @JsonKey()
+  final String hex;
+  @override
+  @JsonKey()
+  final String txId;
+  @override
+  @JsonKey()
+  final String errFileImport;
+  @override
+  @JsonKey()
+  final bool clearPsbt;
+  @override
+  final String? importedPsbtPath;
+  @override
+  final String? importedPsbtfileName;
+  @override
+  final String? importedHexPath;
+  @override
+  final String? importedHexfileName;
+
+  @override
+  String toString({DiagnosticLevel minLevel = DiagnosticLevel.info}) {
+    return 'BroadcastState(broadcasting: $broadcasting, errBroadcasting: $errBroadcasting, psbt: $psbt, hex: $hex, txId: $txId, errFileImport: $errFileImport, clearPsbt: $clearPsbt, importedPsbtPath: $importedPsbtPath, importedPsbtfileName: $importedPsbtfileName, importedHexPath: $importedHexPath, importedHexfileName: $importedHexfileName)';
+  }
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties
+      ..add(DiagnosticsProperty('type', 'BroadcastState'))
+      ..add(DiagnosticsProperty('broadcasting', broadcasting))
+      ..add(DiagnosticsProperty('errBroadcasting', errBroadcasting))
+      ..add(DiagnosticsProperty('psbt', psbt))
+      ..add(DiagnosticsProperty('hex', hex))
+      ..add(DiagnosticsProperty('txId', txId))
+      ..add(DiagnosticsProperty('errFileImport', errFileImport))
+      ..add(DiagnosticsProperty('clearPsbt', clearPsbt))
+      ..add(DiagnosticsProperty('importedPsbtPath', importedPsbtPath))
+      ..add(DiagnosticsProperty('importedPsbtfileName', importedPsbtfileName))
+      ..add(DiagnosticsProperty('importedHexPath', importedHexPath))
+      ..add(DiagnosticsProperty('importedHexfileName', importedHexfileName));
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$_BroadcastState &&
+            const DeepCollectionEquality()
+                .equals(other.broadcasting, broadcasting) &&
+            const DeepCollectionEquality()
+                .equals(other.errBroadcasting, errBroadcasting) &&
+            const DeepCollectionEquality().equals(other.psbt, psbt) &&
+            const DeepCollectionEquality().equals(other.hex, hex) &&
+            const DeepCollectionEquality().equals(other.txId, txId) &&
+            const DeepCollectionEquality()
+                .equals(other.errFileImport, errFileImport) &&
+            const DeepCollectionEquality().equals(other.clearPsbt, clearPsbt) &&
+            const DeepCollectionEquality()
+                .equals(other.importedPsbtPath, importedPsbtPath) &&
+            const DeepCollectionEquality()
+                .equals(other.importedPsbtfileName, importedPsbtfileName) &&
+            const DeepCollectionEquality()
+                .equals(other.importedHexPath, importedHexPath) &&
+            const DeepCollectionEquality()
+                .equals(other.importedHexfileName, importedHexfileName));
+  }
+
+  @override
+  int get hashCode => Object.hash(
+      runtimeType,
+      const DeepCollectionEquality().hash(broadcasting),
+      const DeepCollectionEquality().hash(errBroadcasting),
+      const DeepCollectionEquality().hash(psbt),
+      const DeepCollectionEquality().hash(hex),
+      const DeepCollectionEquality().hash(txId),
+      const DeepCollectionEquality().hash(errFileImport),
+      const DeepCollectionEquality().hash(clearPsbt),
+      const DeepCollectionEquality().hash(importedPsbtPath),
+      const DeepCollectionEquality().hash(importedPsbtfileName),
+      const DeepCollectionEquality().hash(importedHexPath),
+      const DeepCollectionEquality().hash(importedHexfileName));
+
+  @JsonKey(ignore: true)
+  @override
+  _$$_BroadcastStateCopyWith<_$_BroadcastState> get copyWith =>
+      __$$_BroadcastStateCopyWithImpl<_$_BroadcastState>(this, _$identity);
+}
+
+abstract class _BroadcastState extends BroadcastState {
+  const factory _BroadcastState(
+      {final bool broadcasting,
+      final String errBroadcasting,
+      final String psbt,
+      final String hex,
+      final String txId,
+      final String errFileImport,
+      final bool clearPsbt,
+      final String? importedPsbtPath,
+      final String? importedPsbtfileName,
+      final String? importedHexPath,
+      final String? importedHexfileName}) = _$_BroadcastState;
+  const _BroadcastState._() : super._();
+
+  @override
+  bool get broadcasting => throw _privateConstructorUsedError;
+  @override
+  String get errBroadcasting => throw _privateConstructorUsedError;
+  @override
+  String get psbt => throw _privateConstructorUsedError;
+  @override
+  String get hex => throw _privateConstructorUsedError;
+  @override
+  String get txId => throw _privateConstructorUsedError;
+  @override
+  String get errFileImport => throw _privateConstructorUsedError;
+  @override
+  bool get clearPsbt => throw _privateConstructorUsedError;
+  @override
+  String? get importedPsbtPath => throw _privateConstructorUsedError;
+  @override
+  String? get importedPsbtfileName => throw _privateConstructorUsedError;
+  @override
+  String? get importedHexPath => throw _privateConstructorUsedError;
+  @override
+  String? get importedHexfileName => throw _privateConstructorUsedError;
+  @override
+  @JsonKey(ignore: true)
+  _$$_BroadcastStateCopyWith<_$_BroadcastState> get copyWith =>
+      throw _privateConstructorUsedError;
+}
